+++ BE/src/main/java/com/chibbol/wtz/domain/vote/repository/VoteRedisRepository.java
@@ -20,14 +20,14 @@ public class VoteRedisRepository {
     private final RedisTemplate<String, String> redisTemplate;
     private final ObjectMapper objectMapper;
 
-    public List<Vote> findAllByRoomSeqAndTurn(Long roomSeq, Long turn) {
+    public List<Vote> findAllByRoomSeqAndTurn(Long roomSeq, int turn) {
         String key = generateKey(roomSeq, turn);
         List<Object> jsonDataList = redisTemplate.opsForHash().values(key);
         return convertJsonDataListToVoteList(jsonDataList);
     }
 
     public void deleteAllByRoomSeq(Long roomSeq) {
-        Set<String> keys = redisTemplate.keys(generateKey(roomSeq, -1L));
+        Set<String> keys = redisTemplate.keys(generateKey(roomSeq, -1));
         if(keys != null) {
             redisTemplate.delete(keys);
         }
@@ -35,7 +35,7 @@ public class VoteRedisRepository {
 
     public void save(Vote vote) {
         Long roomSeq = vote.getRoomSeq();
-        Long turn = vote.getTurn();
+        int turn = vote.getTurn();
         String key = generateKey(roomSeq, turn);
         String userSeqField = vote.getUserSeq().toString();
 
@@ -64,7 +64,7 @@ public class VoteRedisRepository {
         return resultList;
     }
 
-    private String generateKey(Long roomSeq, Long turn) {
+    private String generateKey(Long roomSeq, int turn) {
         if(turn == -1L) {
             return KEY_PREFIX + ":room:" + roomSeq + ":turn:*";
         }
